requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cycler==0.10.0
easyocr==1.4.1
Flask==2.0.2
idna==3.3
imageio==2.9.0
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.2
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
Pillow==8.2.0
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
requests==2.26.0
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
tifffile==2021.10.12
torch==1.9.1
torchvision==0.10.1
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
